+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactArctangentNodeable.cpp
