+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactCosineNodeable.ScriptCanvasNodeable.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<ScriptCanvas Include="Include/ScriptCanvas/Libraries/Compact/Triginometry/CompactCosineNodeable.h" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<ScriptCanvas Include="Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactCosineNodeable.h" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 			  xsi:noNamespaceSchemaLocation="../../../AutoGen/ScriptCanvasNodeable.xsd">
     <Class Name="CompactCosineNodeable"
         QualifiedName="Nodeables::CompactCosineNodeable"
         PreferredClassName="cos"
-        Category="Compact/Triginometry"
+        Category="Compact/Trigonometry"
         Namespace="ScriptCanvas"
         Description="Gets the cosine of the input number in radians"
         Preset="compact">
