+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactCosineNodeable.cpp
