+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactSineNodeable.ScriptCanvasNodeable.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 
-<ScriptCanvas Include="Include/ScriptCanvas/Libraries/Compact/Triginometry/CompactSineNodeable.h" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+<ScriptCanvas Include="Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactSineNodeable.h" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 			  xsi:noNamespaceSchemaLocation="../../../AutoGen/ScriptCanvasNodeable.xsd">
     <Class Name="CompactSineNodeable"
         QualifiedName="Nodeables::CompactSineNodeable"
         PreferredClassName="sin"
-        Category="Compact/Triginometry"
+        Category="Compact/Trigonometry"
         Namespace="ScriptCanvas"
         Description="Gets the sine of the input number in radians"
         Preset="compact">
