+++ Gems/ScriptCanvas/Code/Include/ScriptCanvas/Libraries/Compact/Trigonometry/CompactTangentNodeable.cpp
