+++ Gems/ScriptCanvasTesting/Assets/ScriptCanvas/UnitTests/LY_SC_UnitTest_CompactNodeTrigonometry.scriptcanvas
